Drop Linux DTS patch in favor of the lmp-device-tree recipe
To support an easier workflow for making changes to custom HW DTS, let's
use the lmp-device-tree recipe instead of managing DTS patch.

This has several advantages:
- Is not tied to a specific Linux recipe
- Easier to update in standard editor
- Easy diff view of historical changes

The lmp-device-tree recipe automatically adds any DTB generated there
to the fit-image for the kernel.  However, we add special changes
so that the lmp-base build also includes these custom DTBs.

NOTE: The lmp-base build is a non-OTA version of the LmP which
follows the standard format of keeping kernel binaries and DTBs
in the vfat partition.  This is useful for HW enablement etc.

Signed-off-by: Michael Scott <mike@foundries.io>